+++ src/main/java/marketplace/entity/Product.java
@@ -16,7 +16,7 @@ public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(name = "title")
     private String title;
@@ -25,7 +25,7 @@ public class Product {
     private BigDecimal price;
 
     @Column(name = "quantity")
-    private int quantity;
+    private Long quantity;
 
     @ManyToOne(cascade = CascadeType.PERSIST)
     @JoinColumn(name = "category_id", nullable = false)
