+++ src/main/java/marketplace/repositories/OrderProductRepository.java
@@ -3,5 +3,5 @@ package marketplace.repositories;
 import marketplace.entity.OrderProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderProductRepository extends JpaRepository<OrderProduct, Integer> {
+public interface OrderProductRepository extends JpaRepository<OrderProduct, Long> {
 }
