+++ src/main/java/marketplace/repositories/ProductRepository.java
@@ -10,11 +10,11 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 
     @Lock(LockModeType.PESSIMISTIC_WRITE)
     @Query("select p from Product p where p.id = :id")
-    public Optional<Product> findByIdForUpdate(int id);
+    public Optional<Product> findByIdForUpdate(Long id);
 }
